GetReviewGroups doesnt use explicit password
Previously, the username and password were passed in to be used as
command line arguments as plain text strings.

This caused them to be logged as plain text in the logging files and
also caused an application exception when the password contained spaces.

Now, it follows the same course of action as request review, where
username and password are omitted.